Update price sync

+++ orm/document/stats_asset.go
@@ -14,9 +14,9 @@ const (
 
 // StatAssetInfoList1H defines the schema for asset statistics in an hourly basis
 type StatAssetInfoList20Minute struct {
-	Price     float32   `bson:"price"`
-	Marketcap float32   `bson:"market_cap"`
-	Volume24H float32   `bson:"volumne_24h"`
+	Price     float64   `bson:"price"`
+	Marketcap float64   `bson:"market_cap"`
+	Volume24H float64   `bson:"volumne_24h"`
 	Timestamp time.Time `bson:"timestamp"`
 }
